### 2025-02-19 #688 @vincenzocoia
distionary: Create and Evaluate Probability Distributions
Submitting Author Name: Vincenzo Coia
Submitting Author Github Handle: @vincenzocoia
Repository:  https://github.com/probaverse/distionary
Version submitted: 0.1.0
Submission type: Stats
Badge grade: silver

Version accepted: TBD
---

```
Package: distionary
Title: Create and Evaluate Probability Distributions
Version: 0.1.0
    person("Vincenzo", "Coia", , "vincenzo.coia@gmail.com",
           role = c("aut", "cre")),
    person("Amogh", "Joshi", role = "ctb"),
    person("Shuyi", "Tan", role = "ctb"),
    person("Zhipeng", "Zhu", role = "ctb")
    )
Description: Create probability distribution objects from a list of
    common distribution families, or make your own distribution.
    Evaluate distribution properties, even if it is not specified
    in the distribution's definition.
    covr,
    tibble
    ellipsis,
    vctrs
```

- Please indicate which of our [statistical package categories](https://stats-devguide.ropensci.org/overview.html#overview-categories) this package falls under. (Please check one or more appropriate boxes below):

- [ ] A pre-submission inquiry has been approved in [issue#<issue_num>](#<issue_num>)

> -   Who is the target audience and what are scientific applications of this package?

Lots of people work with probability distributions. Lots of people don’t work with probability distributions but should, because they don’t see the value or because distributions are too clumsy to work with under existing infrastructure. And, there are lots of people learning about probability distributions that would have an easier time if they get to “feel” distributions and their multifaceted nature. `distionary` is for all of these people.

More accurately, the [_probaverse_](https://probaverse.com) is for all of these people, and `distionary` powers the probaverse. The probaverse is a suite of packages providing an intuitive API for easily making distributions that are representative of your data, noting that typical treatments of distributions are usually not enough for advanced probabilistic modelling.

> -   Paste your responses to our [*General Standard* **G1.1** here](https://stats-devguide.ropensci.org/standards.html#general-standards), describing whether your software is:
>
>     - *The first implementation of a novel algorithm*; or
>     - *The first implementation within* **R** *of an algorithm which has previously been implemented in other languages or contexts*; or
>     - *An improvement on other implementations of similar algorithms in* **R**.
>
>     Please include hyperlinked references to all other relevant software.

I'd say distionary is *An improvement on other implementations of similar algorithms in* **R**.  Lots of software exists for building and evaluating probability distributions, but I've always found the software packages to lack features I need for my work, and packages having different strengths don't talk to each other to form a cohesive API. Over the past 5 years, I developed distionary and the wider probaverse on an as-needed basis so I could more easily make and evaluate distributions for my work. I secured funding from the R Consortium so that I can make `distionary` complete and published (so that the rest of the probaverse can be launched).

Here are some example packages that already exist that handle distributions. I'll only list three, but there are lots.

- [`distributional`](https://pkg.mitchelloharawild.com/distributional/) is a similar package that began development around the same time as the probaverse (back then it was just `distplyr`, for which `distionary` was a part of). We even converged on similar syntax unknowingly. `distributional` has since expanded from its more limited earlier days, but its scope is still limited compared to the goals of the probaverse (simple examples include not being able to make your own distribution families, and not enough operations are available, at least for my applications). Also, `distributional` vectorises distributions, which conflicts with the probaverse philosophy of vectorising evaluation arguments (like CDF arguments) first and letting the user be more deliberate if wanting to also work with a vector of distributions.
- [`distributions3`](https://alexpghayes.github.io/distributions3/) is similar to distributional but looks to be rather limited. Similar comments apply.
- [`fitdistrplus`](https://cran.r-project.org/web/packages/fitdistrplus/vignettes/fitdistrplus_vignette.html) focusses on estimating distribution parameters and is lighter on the distribution specification and evaluation scope. It doesn't talk to other packages like `distributional` or `distributions3` to leverage their strengths.

> -   (If applicable) Does your package comply with our [guidance around *Ethics, Data Privacy and Human Subjects Research*](https://devguide.ropensci.org/policies.html#ethics-data-privacy-and-human-subjects-research)?

This is only tangentially relevant but these practices have been considered in the development of distionary.

> -    What grade of badge are you aiming for? ([bronze, silver, gold](https://stats-devguide.ropensci.org/pkgdev.html#pkgdev-badges))

I think for now, silver. I have lots of plans for building more functionality into this package: improving algorithms, improving how properties are calculated when not specified, specifying _family_ objects in addition to distributions, distributions beyond univariate, reparameterisation, etc., and I already have a design laid out for most of these scopes. Right now, my objective is to get a basic version of distionary reviewed and on CRAN, as a basis for getting the rest of the probaverse up. For instance, I'm OK with having more onus on the user to make sure inputs are correct at this stage. Eventually I do plan to seek a gold badge, because I want to ensure quality as the probaverse develops.

> -    If aiming for silver or gold, describe which of the [four aspects listed in the *Guide for Authors* chapter](https://stats-devguide.ropensci.org/pkgdev.html#pkgdev-silver) the package fulfils (at least one aspect for silver; three for gold)

I'd say the one component extending beyond the minimal requirements is `distionary`'s generality to multiple use cases. I use it for statistical hydrology and climate modelling; I have also used it when putting together a business proposal to simulate projected profit. Anyone doing risk analysis, such as in the actuarial sciences, would find the probaverse useful, too. Financial companies would find it useful for simulating and evaluating uncertainty. Epidemiologists can use it to model the spread of disease. I foresee lots of exciting applications.

- 👀 I/we have run [`autotest`](https://github.com/ropensci-review-tools/autotest) checks on the package, and ensured no tests fail.
    - I suspect this test passes, but I also think there's a bug in the package. I get the error `Error in tools::Rd_db(gsub("^package:", "", i)) : installed help of package ‘distionary’ is corrupt`, but the help files all seem fine. Maybe you won't get the error.
- 👀 The [`srr_stats_pre_submit()` function](https://ropensci-review-tools.github.io/srr/reference/srr_stats_pre_submit.html) confirms this package may be submitted.
    - All standards are present and have been appropriately handled, but I think there's a bug with this package, too. After running `srr_stats_pre_submit()` I get the message `! This package has no 'srr' standards`, despite following the directions. Maybe you won't get the error.
