### 2025-02-19 #688 @mpadge
Thanks for the submission @vincenzocoia. I trust you will have seen the statement in [our stats-devguide](https://stats-devguide.ropensci.org/standards.html#standards-distributions) regarding *Probability Disitributions* software that,

> Unlike most other categories of standards, packages which fit in this category will also generally be expected to fit into at least one other category of statistical software. Reflecting that expectation, standards for probability distributions will be expected to only pertain to some (potentially small) portion of code in any package.

I also immediately note that "distionary" truly is a pure probability distributions package. Given that, I'm confident that we would make an exception here, but would nevertheless ask you first to consider whether our standards for [*Exploratory Data Analysis and Summary Statistics*](https://stats-devguide.ropensci.org/standards.html#standards-eda) might also apply? "distionary" is also very much about providing summary statistics, and it seems to me that at least half of those standards might also apply. Could you please have a quick look and provide feedback on:

1. Whether you think at least 50% of those standards might be applicable? and
2. Whether complying with those standards might improve your package?

Thanks!

-----

Separately, I can't reproduce your {srr} issue, and the bot checks above seem to find all standards okay. I can, however, reproduce your {autotest} issue - can you please file and issue [over there](https://github.com/ropensci-review-tools/autotest/issues), and we'll investigate further. Thanks
