### 2025-02-19 #688 @vincenzocoia
Hi @mpadge, I did see that statement in the stats-devguide, but I didn't think it was relevant for distionary. Thanks for suggesting I look at the [_Exploratory Data Analysis and Summary Statistics_](https://stats-devguide.ropensci.org/standards.html#standards-eda) section.

> 1. Whether you think at least 50% of those standards might be applicable? and

No, I see 35% of these standards being applicable.

> 2. Whether complying with those standards might improve your package?

The standards **EA4.2** and **EA6.0** would improve the package. I don't think the others would.
