### 2025-02-20 #688 @mpadge
Thanks @vincenzocoia, 35% means those standards shouldn't be considered applicable. Feel free regardless to incorporate ideas from EA4.2 and EA6.0, but please don't put any {srr} tags on those. We'll then be happy to consider it a probability distributions package alone. But then we do need to address the requirement that stats packages comply with a minimum of 50% of all standards. {distionary} currently complies with only 26 / 82, mainly through complying with only 15 out of 68 general standards. Many of your `srrstatsNA` statements could actually be compliance statements, notably i ncluding:

- G1.0 - we require that that be complied with regardless, and your case simply explaining absence of primary reference would suffice, maybe with an additional note that (as far as I can tell) all other CRAN packages for distributional representations also lack primary citations.
- G1.3 Again, just converting to `srrstats` with current statement is sufficient to comply

Many other of your `srrstatsNA` tags could and arguably should be complied with, for example:

- G2.1, G2.2
- G2.3, which is directly applicable, for example in `fn_prefix` or `sep` parameters
- G2.4

... and many others. In more general terms:

- Simply implementing more extensive assertions on input data would enable many more to be complied with. [This blog post](https://blog.r-hub.io/2022/03/10/input-checking/) has helpful advice. I personally use [the {checkmate} package](https://mllg.github.io/checkmate/), with which each input can be checked, matched, or asserted with a single additional line of code.
- Missing data, `NA` values, and control of error responses should be included.
- Parameter recover tests are indeed valid, as many distributional specifiers effectively convert inputs into parameters of a distribution, and distributional evaluators are a form of "parameter recovery". [This test in `test-eval_quantile_from_network.R`](https://github.com/probaverse/distionary/blob/main/tests/testthat/test-eval_quantile_from_network.R) is effectively a parameter recovery test. Complying with these standards would then likely just require extending that kind of test to the full range of `eval_` functions.
- Noise susceptibility tests could also readily be implemented, and those standards complied with.

I appreciate that this may require quite a bit of work on your part, but the minimal 50% compliance is a hard requirement from our side, so we ask that you at least go through your current [list of N/A standards](https://github.com/probaverse/distionary/blob/main/R/srr-stats-standards.R) and try to comply with at least 16 more of those, to ensure > 50% in total. Please let us know whether you think that will be possible.
