### 2025-02-20 #688 @maurolepore
@vincenzocoia thanks for sharing your work with rOpenSci, and thanks @mpadge for discussing scope.

I'll share what I quickly see, so as to support the discussion with the editorial board.

@vincenzocoia if any of this helps as feedback, it's a good time to make changes.

### Preliminary checks:

        > It's clear but not grouped. The names sort well but grouping by topic would be nice.
    > Comments by @mpadge suggest it's a good fit but the specific category may need to be defined.
    > ! NEWS.md suggests it's on CRAN but isn't.
    > R/f.R isn't alligned with tests/testthat/test-f.R but test-*.R file groups similar functions meaningfully. Maybe the same groups could also help structure the Reference section of the website?
    > Note the license says 2021. If you submit to CRAN they may requrie the year to be
    updated to the year of submission.
    > I see a few issues labeled with [bug] that have remained opened 1+ years.
