### 2025-02-23 27391b0
rm CRAN submission from NEWS -- that's not ready yet
