+++ NEWS.md
@@ -1,3 +1,2 @@
 # distionary 0.1.0
 
-* Initial CRAN submission.
