### 2025-02-24 aa5c3c0
add structure to reference page in pkgdown
