+++ man/bracket_parameters.Rd
@@ -1,28 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/pretty_name.R
-\name{bracket_parameters}
-\alias{bracket_parameters}
-\title{Present parameters using bracket notation}
-\usage{
-bracket_parameters(distribution, param_digits = 2)
-}
-\arguments{
-\item{distribution}{A single distplyr distribution for which the
-\code{parameters()} function can be applied.}
-
-\item{param_digits}{Number of significant digits to include in the
-output.}
-}
-\value{
-a single character with the parameters separated by commas and
-sandwiched by brackets.
-}
-\description{
-For distributions that have parameters, this function places those
-parameters in bracket notation, such as (0, 1) for a Normal distribution
-with a mean of 0 and variance of 1.
-}
-\examples{
-d <- dst_norm(0, 1)
-distionary:::bracket_parameters(d)
-}
