+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ~
 articles:
   evaluate: evaluate.html
   specify: specify.html
-last_built: 2025-02-24T04:23Z
+last_built: 2025-02-24T04:38Z
 urls:
   reference: https://distionary.probaverse.com/reference
   article: https://distionary.probaverse.com/articles
