### 2025-02-24 6610f60
use checkmate pkg for checking dst_ inputs
