+++ R/dst_gpd.R
@@ -24,14 +24,10 @@
 #' mean(light)
 #' @export
 dst_gpd <- function(scale, shape) {
+  checkmate::assert_numeric(scale, 0, len = 1)
+  checkmate::assert_numeric(shape, len = 1)
   if (is.na(scale) || is.na(shape)) {
     return(dst_null())
-  }
-  if (length(scale) != 1 || length(shape) != 1) {
-    stop("Input parameters must have length 1.")
-  }
-  if (scale <= 0) {
-    stop("'scale' parameter must be positive.")
   }
   distribution(
     parameters = list(scale = scale, shape = shape),
