+++ R/dst_geom.R
@@ -13,14 +13,9 @@
 #' range(d)
 #' @export
 dst_geom <- function(prob) {
+  checkmate::assert_numeric(prob, 0, 1, len = 1)
   if (is.na(prob)) {
     return(dst_null())
-  }
-  if (length(prob) != 1) {
-    stop("Input parameters must have length 1.")
-  }
-  if (prob <= 0 || prob > 1) {
-    stop("prob must be greater than 0 and less than or equal to 1.")
   }
   distribution(
     parameters = list(prob = prob),
