+++ R/dst_exp.R
@@ -8,14 +8,9 @@
 #' dst_exp(1)
 #' @export
 dst_exp <- function(rate) {
+  checkmate::assert_numeric(rate, 0, len = 1)
   if (is.na(rate)) {
     return(dst_null())
-  }
-  if (length(rate) != 1) {
-    stop("Input parameters must have length 1.")
-  }
-  if (rate <= 0) {
-    stop("rate must be greater than 0.")
   }
   distribution(
     parameters = list(rate = rate),
