### 2025-02-24 428f351
stricter control on input data
