+++ R/dst_gev.R
@@ -58,7 +58,10 @@ dst_gev <- function(location, scale, shape) {
         scale^2 * (gamma(1 - 2 * shape) - gamma(1 - shape)^2) / shape^2
       }
     },
-    range = gev_range(location, scale, shape),
+    range = c(
+      gev_lower(location, scale, shape),
+      gev_upper(location, scale, shape)
+    ),
     .name = "Generalised Extreme Value",
     .vtype = "continuous"
   )
