### 2025-02-24 3a64e4f
Vectorize p/d/q functions for GEV and GPD; stricter control on inputs
