+++ R/pdqr_gpd.R
@@ -2,94 +2,100 @@
 #'
 #' @param x,q Vector of quantiles.
 #' @param p Vector of probabilities.
-#' @param location Location parameter to shift the GPD by.
-#' @param scale Scale parameter; single positive numeric.
-#' @param shape Shape parameter; single positive numeric.
-#' @param lower.tail Logical. If `TRUE`, cdf (default);
+#' @param scale Vector of scale parameters; positive numeric.
+#' @param shape Vector of shape parameters; positive numeric.
+#' @param lower.tail Single logical. If `TRUE`, cdf (default);
 #' if `FALSE`, survival function.
-#' @returns Vector of evaluated GPD distribution.
+#' @returns Vector of evaluated GPD distribution, with length
+#' equal to the recycled lengths of `q`/`x`/`p`, `scale`, and `shape`.
 #' @examples
-#' pgpd(1:10, 0, 1, 1)
-#' dgpd(1:10, 0, 2, 0)
-#' qgpd(1:9 / 10, 2, 10, -2)
+#' pgpd(1:10, 1, 1)
+#' dgpd(1:10, 2, 0)
+#' qgpd(1:9 / 10, 10, -2)
 #' @rdname gpd_raw
 #' @export
-pgpd <- function(q, location, scale, shape, lower.tail = TRUE) {
-  if (shape == 0) {
-    left <- q < location
-    z <- (q - location) / scale
-    if (lower.tail) {
-      res <- 1 - exp(-z)
-      res[left] <- 0
-    } else {
-      res <- exp(-z)
-      res[left] <- 1
-    }
-    res
+pgpd <- function(q, scale, shape, lower.tail = TRUE) {
+  checkmate::assert_numeric(q)
+  checkmate::assert_numeric(scale)
+  checkmate::assert_numeric(shape)
+  checkmate::assert_logical(lower.tail, len = 1, any.missing = FALSE)
+  l <- vctrs::vec_recycle_common(q, scale, shape)
+  q <- l[[1]]
+  scale <- l[[2]]
+  shape <- l[[3]]
+  hi <- gpd_upper(scale = scale, shape = shape)
+  left <- q < 0
+  right <- q > hi
+  t <- gev_t_function(q, location = 0, scale = scale, shape = shape)
+  if (lower.tail) {
+    res <- 1 - t
+    res[left] <- 0
+    res[right] <- 1
   } else {
-    if (shape > 0) {
-      rightend <- Inf
-    } else {
-      rightend <- location - scale / shape
-    }
-    left <- q < location
-    right <- q > rightend
-    z <- (q - location) / scale
-    if (lower.tail) {
-      res <- 1 - (1 + shape * z)^(-1 / shape)
-      res[left] <- 0
-      res[right] <- 1
-    } else {
-      res <- (1 + shape * z)^(-1 / shape)
-      res[left] <- 1
-      res[right] <- 0
-    }
-    res
+    res <- t
+    res[left] <- 1
+    res[right] <- 0
   }
+  res
 }
 
 
 #' @rdname gpd_raw
-#' @inheritParams pgpd
 #' @export
-qgpd <- function(p, location, scale, shape) {
-  invalid <- p < 0 | p > 1
-  if (shape == 0) {
-    res <- location - scale * log(1 - p)
-    res[invalid] <- NaN
-    res
-  } else {
-    if (shape > 0) {
-      rightend <- Inf
-    } else {
-      rightend <- location - scale / shape
-    }
-    t <- 1 / (1 - p)
-    res <- location + scale * (t^shape - 1) / shape
[…]
