+++ R/pdqr_gev.R
@@ -2,83 +2,136 @@
 #'
 #' @param x,q Vector of quantiles.
 #' @param p Vector of probabilities.
-#' @param location Location parameter; single numeric.
-#' @param scale Scale parameter; single positive numeric.
-#' @param shape Shape parameter; single numeric.
+#' @param location Location parameter; numeric vector.
+#' @param scale Scale parameter; positive numeric vector.
+#' @param shape Shape parameter; numeric vector.
 #' This is also the extreme value index,
 #' so that `shape > 0` is heavy tailed, and `shape < 0` is short-tailed.
-#' @returns Vector of evaluated GEV distribution.
+#' @returns Vector of evaluated GEV distribution, with length
+#' equal to the recycled lengths of `q`/`x`/`p`, `location`, `scale`, and
+#' `shape`.
 #' @examples
 #' pgev(1:10, 0, 1, 1)
-#' dgev(1:10, 0, 2, 0)
+#' dgev(1:10, 1:10, 2, 0)
 #' qgev(1:9 / 10, 2, 10, -2)
 #' @rdname gev_raw
 #' @export
 pgev <- function(q, location, scale, shape) {
-  r <- gev_range(location, scale, shape)
+  checkmate::assert_numeric(q)
+  checkmate::assert_numeric(location)
+  checkmate::assert_numeric(scale)
+  checkmate::assert_numeric(shape)
+  l <- vctrs::vec_recycle_common(q, location, scale, shape)
+  q <- l[[1]]
+  location <- l[[2]]
+  scale <- l[[3]]
+  shape <- l[[4]]
+  lo <- gev_lower(location, scale, shape)
+  hi <- gev_upper(location, scale, shape)
   t <- gev_t_function(q, location = location, scale = scale, shape = shape)
   res <- exp(-t)
-  res[q <= r[1L]] <- 0
-  res[q >= r[2L]] <- 1
+  res[q <= lo] <- 0
+  res[q >= hi] <- 1
   res
 }
 
 #' @rdname gev_raw
-#' @inheritParams pgev
 #' @export
 qgev <- function(p, location, scale, shape) {
-  invalid_at <- p < 0 | p > 1
-  if (shape == 0) {
-    res <- location - scale * log(-log(p))
-  } else {
-    res <- location + scale * ((-log(p))^(-shape) - 1) / shape
-  }
-  res[invalid_at] <- NaN
-  res
+  checkmate::assert_numeric(p)
+  checkmate::assert_numeric(location)
+  checkmate::assert_numeric(scale)
+  checkmate::assert_numeric(shape)
+  l <- vctrs::vec_recycle_common(p, location, scale, shape)
+  p <- l[[1]]
+  location <- l[[2]]
+  scale <- l[[3]]
+  shape <- l[[4]]
+  i_zero <- shape == 0
+  i_non0 <- shape != 0
+  res <- numeric(length(p))
+  neglogp <- -log(p)
+  res[i_zero] <- -log(neglogp[i_zero])
+  res[i_non0] <- (neglogp[i_non0]^(-shape[i_non0]) - 1) / shape[i_non0]
+  res[p < 0 | p > 1] <- NaN
+  location + scale * res
 }
 
 #' @rdname gev_raw
-#' @inheritParams pgev
 #' @export
 dgev <- function(x, location, scale, shape) {
-  r <- gev_range(location, scale, shape)
+  checkmate::assert_numeric(x)
+  checkmate::assert_numeric(location)
+  checkmate::assert_numeric(scale)
+  checkmate::assert_numeric(shape)
+  l <- vctrs::vec_recycle_common(x, location, scale, shape)
+  x <- l[[1]]
+  location <- l[[2]]
+  scale <- l[[3]]
+  shape <- l[[4]]
[…]
