### 2025-02-24 9f773e2
finish checking inputs
