+++ R/representation_as_function.R
@@ -2,10 +2,12 @@
 #'
 #' @param distribution Distribution to extract a representation from.
 #' @param representation Character, such as `"cdf"`. In general, a suffix
-#' to an `eval_` function.
+#' to an `eval_` function. Vector of length 1.
 #' @returns A function of the representation.
 #' @noRd
 representation_as_function <- function(distribution, representation) {
+  checkmate::assert_class(distribution, "dst")
+  checkmate::assert_character(representation, len = 1)
   f <- distribution[[representation]]
   if (is.null(f)) {
     f <- \(x) eval_property(distribution, representation, x)
