+++ R/skewness.R
@@ -1,5 +1,6 @@
 #' @rdname moments
 #' @export
 skewness <- function(distribution) {
+  checkmate::assert_class(distribution, "dst")
   eval_property(distribution, "skewness")
 }
