### 2025-02-25 44f90a4
further tests to vectorized p/d/q GEV and GPD functions
