+++ tests/testthat/test-dst_gpd.R
@@ -1,58 +1,66 @@
-test_that("GPD quantities work out, shape > 0", {
+test_that("GPD quantities match known cases, shape > 0", {
   scale <- 1
   shape <- 1
-  d <- dst_gpd(scale = scale, shape = shape)
   med <- scale * (2^shape - 1) / shape
-  expect_equal(eval_cdf(d, c(-1, 0, med)), c(0, 0, 0.5))
-  expect_equal(eval_quantile(d, c(0, 0.5, 1)), c(0, med, Inf))
-  expect_equal(eval_density(d, c(-1, Inf)), c(0, 0))
-  expect_true(all(diff(eval_density(d, 1:10)) < 0))
-  pdf <- representation_as_function(d, "density")
+  expect_equal(
+    pgpd(c(-1, 0, med), scale = scale, shape = shape), c(0, 0, 0.5)
+  )
+  expect_equal(
+    qgpd(c(-1, 0, 0.5, 1, 2), scale = scale, shape = shape),
+    c(NaN, 0, med, Inf, NaN)
+  )
+  expect_equal(dgpd(c(-1, Inf), scale = scale, shape = shape), c(0, 0))
+  expect_true(all(diff(dgpd(1:10, scale = scale, shape = shape)) < 0))
+  pdf <- \(x) dgpd(x, scale = scale, shape = shape)
   expect_equal(integrate(pdf, 0, Inf)$value, 1, tolerance = 0.0001)
-  expect_true(all(realise(d, 10) >= 0))
 })
 
 
-test_that("GPD quantities work out, shape = 0", {
+test_that("GPD quantities match known cases, shape = 0", {
   scale <- 1
   shape <- 0
-  d <- dst_gpd(scale = scale, shape = shape)
   med <- log(2)
-  expect_equal(eval_cdf(d, c(-1, 0, med, Inf)), c(0, 0, 0.5, 1))
-  expect_equal(eval_quantile(d, c(0, 0.5, 1)), c(0, med, Inf))
-  expect_equal(eval_density(d, c(-1, Inf)), c(0, 0))
-  pdf <- representation_as_function(d, "density")
+  expect_equal(
+    pgpd(c(-1, 0, med, Inf), scale = scale, shape = shape), c(0, 0, 0.5, 1)
+  )
+  expect_equal(
+    suppressWarnings(qgpd(c(-1, 0, 0.5, 1, 2), scale = scale, shape = shape)),
+    c(NaN, 0, med, Inf, NaN)
+  )
+  expect_equal(dgpd(c(-1, Inf), scale = scale, shape = shape), c(0, 0))
+  pdf <- \(x) dgpd(x, scale = scale, shape = shape)
   expect_true(all(diff(pdf(1:10)) < 0))
   expect_equal(integrate(pdf, 0, Inf)$value, 1, tolerance = 0.0001)
-  expect_true(all(realise(d, 10) >= 0))
 })
 
-test_that("GPD quantities work out, shape < 0", {
+test_that("GPD quantities match known cases, shape < 0", {
   scale <- 2
   shape <- -1
   rightend <- -scale / shape
-  d <- dst_gpd(scale = scale, shape = shape)
   med <- scale * (2^shape - 1) / shape
   expect_equal(
-    eval_cdf(d, c(-1, 0, med, rightend, rightend + 1, Inf)),
+    pgpd(
+      c(-1, 0, med, rightend, rightend + 1, Inf),
+      scale = scale, shape = shape
+    ),
     c(0, 0, 0.5, 1, 1, 1)
   )
[…]
