+++ tests/testthat/test-pdq_gev.R
@@ -1,0 +1,122 @@
+test_that("GEV negative scale not allowed.", {
+  expect_error(dgev(1:10, 3, -4, 1))
+  expect_error(pgev(1:10, 3, -4, 1))
+  expect_error(qgev(1:10 / 10, 3, -4, 1))
+})
+
+test_that("cdf and pdf of GEV align via numerical derivative.", {
+  d <- list(
+    list(location = 0.2, scale = 1.2, shape = 0.8),
+    list(location = 0.2, scale = 1.2, shape = 0),
+    list(location = 0.2, scale = 1.2, shape = -0.8)
+  )
+  x <- -5:5
+  eps <- 1e-6
+  for (i in seq_along(d)) {
+    pdf <- rlang::exec(dgev, x, !!!d[[i]])
+    cdf1 <- rlang::exec(pgev, x - eps, !!!d[[i]])
+    cdf2 <- rlang::exec(pgev, x, !!!d[[i]])
+    pdf_num <- (cdf2 - cdf1) / eps
+    expect_equal(pdf, pdf_num, tolerance = 1e-6)
+  }
+})
+
+test_that("cdf and qf of GEV align.", {
+  d <- list(
+    list(location = 0.2, scale = 1.2, shape = 0.8),
+    list(location = 0.2, scale = 1.2, shape = 0),
+    list(location = 0.2, scale = 1.2, shape = -0.8)
+  )
+  p <- 1:9 / 10
+  for (i in seq_along(d)) {
+    qf <- rlang::exec(qgev, p, !!!d[[i]])
+    cdf <- rlang::exec(pgev, qf, !!!d[[i]])
+    expect_equal(cdf, p)
+  }
+})
+
+test_that("quantile function of GEV is valid, validating the distribution.", {
+  d <- list(
+    list(location = 0.2, scale = 1.2, shape = 0.8),
+    list(location = 0.2, scale = 1.2, shape = 0),
+    list(location = 0.2, scale = 1.2, shape = -0.8)
+  )
+  p <- 0:100 / 100
+  for (i in seq_along(d)) {
+    lo <- rlang::exec(distionary:::gev_lower, !!!d[[i]])
+    hi <- rlang::exec(distionary:::gev_upper, !!!d[[i]])
+    r <- c(lo, hi)
+    qf <- rlang::exec(qgev, p, !!!d[[i]])
+    expect_true(all(diff(qf) > 0))
+    expect_equal(min(qf), lo)
+    expect_equal(max(qf), hi)
+  }
+})
+
+test_that("vectorisation of p/d/q GEV functions works.", {
+  y <- 1:10
+  x <- 1:10
+  x[4] <- NA_real_
+  x[2] <- NaN
+  p <- x / 11
+  v <- y / 11
+  # Mismatched lengths
+  expect_error(pgev(x, 0:1, 1, 1))
+  expect_error(pgev(x, 0, 1:2, 1))
+  expect_error(pgev(x, 0, 1, 0:1))
+  expect_error(dgev(x, 0:1, 1, 1))
+  expect_error(dgev(x, 0, 1:2, 1))
+  expect_error(dgev(x, 0, 1, 0:1))
+  expect_error(qgev(p, 0:1, 1, 1))
+  expect_error(qgev(p, 0, 1:2, 1))
+  expect_error(qgev(p, 0, 1, 0:1))
+  expect_error(distionary:::gev_lower(0:10, 1:2, 1))
+  expect_error(distionary:::gev_lower(1, 1:2, 1:4))
[…]

[…]
