+++ man/gpd_raw.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/pdqr_gpd.R
+% Please edit documentation in R/pdq_gpd.R
 \name{pgpd}
 \alias{pgpd}
 \alias{qgpd}
