+++ man/gev_raw.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/pdqr_gev.R
+% Please edit documentation in R/pdq_gev.R
 \name{pgev}
 \alias{pgev}
 \alias{qgev}
