+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ~
 articles:
   evaluate: evaluate.html
   specify: specify.html
-last_built: 2025-02-24T08:05Z
+last_built: 2025-02-25T01:52Z
 urls:
   reference: https://distionary.probaverse.com/reference
   article: https://distionary.probaverse.com/articles
