### 2025-03-14 d8591e8
adopt more ROpenSci standards
