+++ R/srr-stats-standards.R
@@ -5,7 +5,62 @@
 #' @srrstats {G1.3} Users are referred to general texts in probability to learn more about probability concepts and terminology. Terms specific to the package (like "distributional representation" are defined in vignettes).
 #' @srrstats {G1.4} Software uses [`roxygen2`](https://roxygen2.r-lib.org/) to document all functions.
 #' @srrstats {G1.4a} All internal (non-exported) functions are documented in standard [`roxygen2`](https://roxygen2.r-lib.org/) format, along with a final `@noRd` tag to suppress automatic generation of `.Rd` files.
-#' @srrstats {PD2.0} Representing probability distributions using a package for general representation is the main purpose of distionary.
+#' @srrstats {G2.0} Assertions on lengths of inputs (asserting that
+#' inputs expected to be single- or multi-valued) are explicitly
+#' tested for distribution parameters; implicitly through evaluation
+#' functions.
+#' @srrstats {G2.1} Assertions on types of inputs is conducted using the checkmate package for most functions.
+#' @srrstats {G2.2} Prohibiting or restricting submission of multivariate input (i.e., distributions) to parameters expected to be univariate is done using the checkmate package for relevant functions (e.g., `dst_*()` specifications)
+#' @srrstats {G2.3} Univariate character input specifications are asserted using the checkmate package where relevant (e.g., `.vtype` and `.name` in `distribution()`; `arg_name` and `fn_prefix` in `enframe_*()`).
+#' @srrstats {G2.4} Mechanisms to convert between different data types is bypassed by requiring strict type inputs (except integer, which is allowed to be integerish).
+#' @srrstats {G2.14} distionary is designed to propagate NA as if it's just another data type.
+#' @srrstats {G2.14a} No option is given to error on missing data; if a user wants this behaviour, it should be explicitly specified in their code, because there is nothing fishy about NA inputs in the distionary context.
+#' @srrstats {G2.14b} NA inputs are "ignored" in the sense that they are not treated as special, but rather just another type of data, and therefore does not need to alert the user of their presence.
+#' @srrstats {G2.14c} Replacing data with imputed data does not make sense in the context of distionary, so is not done.
+#' @srrstats {G2.16} This version of distionary does forces the propagation of undefined values (e.g., `NaN`, `Inf` and `-Inf`) rather than allowing user specification for length-stability, also because `Inf` and `-Inf` are expected in some cases (e.g., the support of any Normal distribution).
+
+#' @srrstats {G5.6} Parameter recovery is relevant when distributional properties (like quantiles) are computed from other properties (like the cdf); these are all tested in `test-distributions_valid.R` when the computational version is compared to the embedded property.
+#' @srrstats {G5.6a} Parameter recovery tests are conducted using a reasonable tolerance.
+#' @srrstats {G2.6} distionary asserts one-dimensional input where required (e.g., `dst_*()` specifications) using the checkmate package.
+#' @srrstats {G3.0} Appropriate tolerances for approximate equality is
+#' adopted (stricter tolerances planned for future based on discretes
+#' tracking design). See `test-distributions_valid.R`.
+#' @srrstats {G5.3} Functions that are expected to return objects containing no missing (`NA`) or undefined (`NaN`, `Inf`) values are tested either implicitly (e.g., `is_distribution()` implicitly checks non-NA value) or explicitly (e.g., `pretty_name()` is never NA).
+#' @srrstats {G5.2} Appropriate error behaviour is tested for all
+#' functions explicitly, but warnings are omitted and saved for a future
+#' version. See `test-distributions_valid.R`.
[…]

[…]
