+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ~
 articles:
   evaluate: evaluate.html
   specify: specify.html
-last_built: 2025-03-14T05:55Z
+last_built: 2025-03-14T05:59Z
 urls:
   reference: https://distionary.probaverse.com/reference
   article: https://distionary.probaverse.com/articles
