### 2025-03-14 135ea0d
Fix Null distribution to have NA type
