### 2025-03-14 f6f0ec0
Update vignette with new .parameters specification.
