+++ vignettes/specify.Rmd
@@ -83,14 +83,13 @@ See the [Evaluate a Distribution] vignette for more details on these evaluation 
 
 For this version of `distionary`, it is mandatory to specify at least `cdf` in order for non-specified properties to be evaluated. For continuous or discrete distributions, `density` or `pmf` may also be needed.
 
-Optionally, you can also specify an entry for `parameters`, which should be a named list of parameters that define the distribution. These parameters are never used for distribution calculations, but are sometimes useful to keep track of for analysis. (A future version of `distionary` will make heavier use of these.)
+Optionally, you can also specify an entry for `.parameters`, which should be a named list of parameters that define the distribution. These parameters are never used for distribution calculations, but are sometimes useful to keep track of for analysis. It is intended that a future version of `distionary` will refer to these parameters for its calculations, but will not break existing code. 
 
 Here is an example distribution.
 
 ```{r}
 # Make a continuous distribution
 linear <- distribution(
-  parameters = list(a = 1),
   density = function(x) {
     d <- 2 * (1 - x)
     d[x < 0 | x > 1] <- 0
@@ -105,7 +104,8 @@ linear <- distribution(
   g = 9.81,
   another_representation = function(x) x^2,
   .vtype = "continuous",
-  .name = "My Linear"
+  .name = "My Linear",
+  .parameters = list(a = 1)
 )
 # Inspect
 linear
