### 2025-03-15 a36eb04
Move big loop of tests into separate files; make internal data file to facilitate
