+++ R/validate_range.R
@@ -1,6 +1,4 @@
 #' @noRd
-#' @inheritParams validate_density
-#' @family validate_properties
 validate_range <- function(distribution, verbose = FALSE, tol = 1e-05) {
   checkmate::assert_class(distribution, "dst")
   checkmate::assert_logical(verbose, len = 1, any.missing = FALSE)
