### 2025-03-16 19bb1ce
Add functionality to get and check intrinsically defined properties
