### 2025-03-21 1db7e59
fix realise() for hypergeometric.
