### 2025-03-21 d56981b
develop formalized test suite for realise()
