### 2025-03-24 231fea2
Force location = 0 for GPD in this temporary version of distionary
