### 2025-03-25 4b27a33
touch up enframe() tests
