### 2025-03-25 a7adda9
fix machine tolerance tests
