+++ tests/testthat/test-machine_tolerance.R
@@ -4,6 +4,7 @@
 #' parameters and evaluation inputs and compared to originals.
 test_that("Distributions are not sensitive to machine tolerance.", {
   for (item in test_distributions) {
+    cat(item$distribution)
     paramset_orig <- item$valid[[1]]
     eps <- .Machine$double.eps
     paramset_eps <- lapply(paramset_orig, \(x) x - eps)
@@ -55,19 +56,40 @@ test_that("Distributions are not sensitive to machine tolerance.", {
       stdev(d_eps)
     )
     ## Skewness
-    expect_equal(
-      skewness(d_orig),
-      skewness(d_eps)
-    )
+    if (pretty_name(d_orig) == "Log Pearson Type III") {
+      expect_equal(
+        suppressMessages(skewness(d_orig)),
+        suppressMessages(skewness(d_eps))
+      )
+    } else {
+      expect_equal(
+        skewness(d_orig),
+        skewness(d_eps)
+      )
+    }
     ## Kurtosis
-    expect_equal(
-      kurtosis(d_orig),
-      kurtosis(d_eps)
-    )
+    if (pretty_name(d_orig) == "Log Pearson Type III") {
+      expect_equal(
+        suppressMessages(kurtosis(d_orig)),
+        suppressMessages(kurtosis(d_eps))
+      )
+    } else {
+      expect_equal(
+        kurtosis(d_orig),
+        kurtosis(d_eps)
+      )
+    }
     ## Excess Kurtosis
-    expect_equal(
-      kurtosis_exc(d_orig),
-      kurtosis_exc(d_eps)
-    )
+    if (pretty_name(d_orig) == "Log Pearson Type III") {
+      expect_equal(
+        suppressMessages(kurtosis_exc(d_orig)),
+        suppressMessages(kurtosis_exc(d_eps))
+      )
+    } else {
+      expect_equal(
+        kurtosis_exc(d_orig),
+        kurtosis_exc(d_eps)
+      )
+    }
   }
 })
