+++ tests/testthat/test-parameters.R
@@ -6,5 +6,11 @@ test_that("Parameters are specified and retrieved correctly w parameters()", {
     d <- rlang::exec(item$distribution, !!!paramset)
     param_retrieved[[i]] <- parameters(d)
     expect_identical(param_retrieved[[!!i]], !!paramset)
+    parameters(d)[[1]] <- "hello"
+    expect_identical(parameters(d)[[1]], "hello")
+    parameters(d)[[1]] <- NaN
+    expect_identical(parameters(d)[[1]], NaN)
+    parameters(d)[[1]] <- data.frame(x = 1)
+    expect_identical(parameters(d)[[1]], data.frame(x = 1))
   }
 })
