### 2025-03-25 652f284
Insist that quantile algo does not accommodate non-continuous distributions yet.
