### 2025-03-25 be424e0
Finalize tests for kurtosis and mean
