+++ tests/testthat/test-property-kurtosis.R
@@ -1,98 +1,75 @@
-test_that("Network is invoked in priority: stdev", {
-  d <- distribution(
-    density = stats::dnorm,
-    range = c(-Inf, Inf),
-    kurtosis = 100, # deliberately incorrect
-    .vtype = "continuous"
-  )
-  # Standard deviation should come from built-in variance.
-  expect_equal(kurtosis_exc(d), 100 - 3)
-  expect_equal(eval_kurtosis_exc_from_network(d), 100 - 3)
-  # Remove variance, and sd of 3 should be retrieved through density.
-  d$kurtosis <- NULL
-  expect_equal(kurtosis_exc(d), 0)
-  expect_equal(eval_kurtosis_exc_from_network(d), 0)
-})
-
-test_that("Network is invoked in priority: variance", {
+test_that("Network is invoked in priority: kurtosis", {
+  # First look for `kurtosis_exc`, then invoke algorithm if not found.
   d <- distribution(
     density = stats::dnorm,
     range = c(-Inf, Inf),
     kurtosis_exc = 100 - 3, # deliberately incorrect
     .vtype = "continuous"
   )
-  # Standard deviation should come from built-in variance.
   expect_equal(kurtosis(d), 100)
   expect_equal(eval_kurtosis_from_network(d), 100)
-  # Remove variance, and sd of 3 should be retrieved through density.
   d$kurtosis_exc <- NULL
   expect_equal(kurtosis(d), 3)
   expect_equal(eval_kurtosis_from_network(d), 3)
 })
 
-test_that("Mean calculated thru network matches known vals", {
+test_that("Original kurtosis is never specified in our distributions", {
+  # ...so don't need to verify their entries.
   for (item in test_distributions) {
-    for (paramset in item$valid) {
-      d <- rlang::exec(item$distribution, !!!paramset)
-      if (is_intrinsic(d, "mean")) {
-        expect_equal(eval_mean_from_network(d), mean(d))
-      }
-    }
+    d <- rlang::exec(item$distribution, !!!item$valid[[1]])
+    expect_false(is_intrinsic(d, "kurtosis"))
   }
 })
 
-test_that("Kurtosis calculated thru network matches known vals", {
+test_that("Kurtosis algorithm matches known vals", {
   for (item in test_distributions) {
     for (paramset in item$valid) {
       d <- rlang::exec(item$distribution, !!!paramset)
-      if (is_intrinsic(d, "kurtosis")) {
-        expect_equal(eval_kurtosis_from_network(d), kurtosis(d))
-        if (is_intrinsic(d, "kurtosis_exc")) {
-          d[["kurtosis_exc"]] <- NULL
-          expect_equal(eval_kurtosis_from_network(d), kurtosis(d))
+      if (is_intrinsic(d, "kurtosis_exc") || is_intrinsic(d, "kurtosis")) {
+        supposed_kurt <- kurtosis(d)
+        if (is.infinite(supposed_kurt)) {
+          supposed_kurt <- NaN  # To align with numerical integration output.
+        }
+        if (vtype(d) == "continuous") {
+          expect_equal(suppressMessages(algorithm_kurtosis(d)), supposed_kurt)
+        } else if (
+          pretty_name(d) %in%
+          c("Hypergeometric", "Bernoulli", "Binomial")
+        ) {
+          expect_error(algorithm_kurtosis(d))
+          r <- range(d)
+          x <- seq(r[1], r[2], by = 1L)
+          mu <- mean(d)
+          sigma <- stdev(d)
+          x4 <- ((x - mu) / sigma)^4
+          p <- eval_pmf(d, at = x)
[…]
