+++ R/eval_from_network-mean.R
@@ -10,14 +10,31 @@ eval_mean_from_network <- function(distribution, ...) {
   dens <- representation_as_function(distribution, representation = "density")
   integrand <- function(x) x * dens(x)
   r <- range(distribution)
+  if (r[1] == -Inf && r[2] == Inf) {
+    # Cauchy distribution mistakenly comes back with finite mean; break
+    # integral into two to solve this issue.
+    int1 <- try(
+      stats::integrate(
+        integrand,
+        lower = r[1], upper = 0,
+        rel.tol = 1e-09,
+        subdivisions = 200L,
+        ...
+      )$value,
+      silent = TRUE
+    )
+    r[1] <- 0
+  } else {
+    int1 <- 0
+  }
   int <- try(
-    stats::integrate(
+    int1 + stats::integrate(
       integrand,
       lower = r[1], upper = r[2],
       rel.tol = 1e-09,
       subdivisions = 200L,
       ...
-    ),
+    )$value,
     silent = TRUE
   )
   if (inherits(int, "try-error")) {
@@ -27,5 +44,5 @@ eval_mean_from_network <- function(distribution, ...) {
     )
     return(NaN)
   }
-  int$value
+  int
 }
