+++ tests/testthat/test-property-skewness.R
@@ -1,37 +1,55 @@
-test_that("Skewness calculated thru network matches known vals", {
+test_that("Skewness algorithm matches known vals", {
   for (item in test_distributions) {
     for (paramset in item$valid) {
       d <- rlang::exec(item$distribution, !!!paramset)
       if (is_intrinsic(d, "skewness")) {
-        expect_equal(eval_skewness_from_network(d), skewness(d))
+        supposed_skew <- skewness(d)
+        if (is.infinite(supposed_skew)) {
+          supposed_skew <- NaN  # To align with numerical integration output.
+        }
+        if (vtype(d) == "continuous") {
+          expect_equal(
+            suppressMessages(eval_skewness_from_network(d)),
+            supposed_skew
+          )
+        } else if (
+          pretty_name(d) %in%
+          c("Hypergeometric", "Bernoulli", "Binomial")
+        ) {
+          expect_error(eval_skewness_from_network(d))
+          r <- range(d)
+          x <- seq(r[1], r[2], by = 1L)
+          mu <- mean(d)
+          sigma <- stdev(d)
+          x3 <- ((x - mu) / sigma)^3
+          p <- eval_pmf(d, at = x)
+          expect_equal(sum(p * x3), supposed_skew)
+        } else if (
+          pretty_name(d) %in%
+          c("Negative Binomial", "Poisson", "Geometric")
+        ) {
+          expect_error(eval_skewness_from_network(d))
+          mu <- mean(d)
+          sigma <- stdev(d)
+          to_add <- Inf
+          i <- 0
+          sk <- 0
+          while (to_add > 1e-9) {
+            x <- 0:99 + 100 * i
+            to_add <- sum(eval_pmf(d, x) * ((x - mu) / sigma)^3)
+            sk <- sk + to_add
+            i <- i + 1
+          }
+          expect_equal(sk, supposed_skew)
+        } else if (pretty_name(d) == "Degenerate") {
+          # Do nothing; checked in a separate script.
+        } else {
+          # Shouldn't be any distributions left.
+          stop("At least one distribution family has not been accounted for.")
+        }
       }
     }
   }
 })
 
-# if (attr(distribution, "name") %in% c(
-#   "Hypergeometric", "Bernoulli", "Binomial"
-# )) {
-#   # This case is for double-checking the moments supplied for these
-#   # distributions, and will be included until discretes handling is
-#   # implemented.
-#   r <- range(distribution)
-#   x <- seq(r[1], r[2], by = 1L)
-#   x3 <- ((x - mu) / sigma)^3
-#   p <- eval_pmf(distribution, at = x)
-#   return(sum(p * x3))
-# }
-# if (attr(distribution, "name") %in% c(
-#   "Negative Binomial", "Poisson", "Geometric"
-# )) {
-#   to_add <- Inf
-#   i <- 0
-#   skew <- 0
-#   while (to_add > 1e-9) {
-#     x <- 0:99 + 100 * i
-#     to_add <- sum(eval_pmf(distribution, x) * ((x - mu) / sigma)^3)
-#     skew <- skew + to_add
-#     i <- i + 1
-#   }
-#   return(skew)
-# }
+
