### 2025-03-29 0b8b5a0
add machine tolerance tests for moments
