+++ tests/testthat/test-property-kurtosis.R
@@ -72,4 +72,9 @@ test_that("Kurtosis algorithm matches known vals", {
   }
 })
 
-
+test_that("Kurtosis algorithm takes longer for smaller tolerance.", {
+  d <- dst_norm(0, 1)
+  t1 <- system.time(algorithm_kurtosis(d, tol = 1e-4))
+  t2 <- system.time(algorithm_kurtosis(d, tol = 1e-8))
+  expect_lt(t1[["elapsed"]], t2[["elapsed"]])
+})
