+++ tests/testthat/test-property-mean.R
@@ -59,3 +59,10 @@ test_that("Mean calculated thru network matches known vals", {
     }
   }
 })
+
+test_that("Mean algorithm takes longer for smaller tolerance.", {
+  d <- dst_norm(0, 1)
+  t1 <- system.time(eval_mean_from_network(d, tol = 1e-4))
+  t2 <- system.time(eval_mean_from_network(d, tol = 1e-8))
+  expect_lt(t1[["elapsed"]], t2[["elapsed"]])
+})
