### 2025-04-01 7b9d67c
switch from integrate() to cubature
