### 2025-04-01 7bdea2d
rm validate_*() functions because they are built-in to the tests
