+++ R/distionary_integrate.R
@@ -1,0 +1,140 @@
+#' Integration regime used in distionary
+#'
+#' Integrating distribution-related functions can be finicky. This function
+#' is designed to get around the shortcomings of built-in integration
+#' regimes like stats::integrate and cubature::hcubature.
+#'
+#' @param fun Function (integrand) to be integrated.
+#' @param lower,upper Limits of integration. Can be infinite.
+#' @param tol Tolerance; small positive number.
+#' @param ... Other arguments to pass to `cubature::hcubature()`.
+#' @returns A single numeric, the integral of the function.
+#' @details Here are the integration shortfalls encountered with distributions:
+#'
+#' - The mean of a standard Cauchy distribution integrates to 0, when it
+#'   should not exist. This happens because the function is symmetric and
+#'   the negative part cancels out the positive part numerically, but
+#'   both sides of the integral are not well-defined.
+#' - Functions that go to infinity as x -> constant sometimes come back as
+#'   infinite, when that's not true (e.g., some Beta distributions).
+#'
+#' The solution involves `if` statements evaluating:
+#'
+#' - If both ends are infinite, cut the integral in half.
+#' - Then, if either end is finite but does not have a finite evaluation,
+#'   change the variable to evaluate to an infinite endpoint.
+#'
+#' **Change of variables math**
+#' Integral is I = int f(x) dx, x = x0..x1
+#'
+#' _Upper Integral_
+#' Let t = 1 / (x1 - x)
+#' ==> x = x1 - 1/t
+#' t goes from 1 / (x1 - x0) .. Inf
+#' dx = 1/t^2 dt
+#' Integral becomes:
+#' I = int f(x1 - 1/t) / t^2 dt
+#'
+#' _Lower Integral_
+#' Let t = 1 / (x - x0)
+#' ==> x = x0 + 1/t
+#' dx = -1/t^2 dt
+#' t goes from Inf .. 1 / (x1 - x0)
+#' Integral becomes
+#' I = int -f(x0 + 1/t) / t^2 dt
+#' Or, switch the order of integration, and get
+#' I = int  f(x0 + 1/t) / t^2 dt
+#'
+#' _Both_
+#' When both finite endpoints have non-finite evaluation, split the
+#' integral in half using the midpoint.
+#'
+#' @examples
+#' # Beta(0.5, 0.5) should have mean of 0.5.
+#' integrand <- \(x) x * stats::dbeta(x, 0.5, 0.5)
+#' plot(integrand)
+#' distionary_integrate(integrand, 0, 1)
+#' stats::integrate(integrand, 0, 1)  # Good
+#' cubature::hcubature(integrand, 0, 1, tol = 1e-7)
+#' cubature::hcubature(integrand, 0, 1)  # Works but not as accurate
+#'
+#' # Cauchy(0, 1)
+#' integrand <- \(x) x * stats::dcauchy(x)
+#' plot(integrand, -10, 10)
+#' distionary_integrate(integrand, 0, Inf)
+#' stats::integrate(integrand, -Inf, Inf)  # 0
+#' cubature::hcubature(integrand, -Inf, Inf) # Correctly NaN
+#' cubature::hcubature(integrand, -Inf, Inf, tol = 1e-9)
+#' @noRd
+distionary_integrate <- function(fun, lower, upper, tol = 1e-9, ...) {
+  checkmate::assert_function(fun)
+  checkmate::assert_numeric(lower, len = 1)
+  checkmate::assert_numeric(upper, lower = lower, len = 1)
+  checkmate::assert_numeric(tol, lower = 0, len = 1)
[…]
