+++ tests/testthat/test-property-variance.R
@@ -74,10 +74,3 @@ test_that("Variance algorithm matches known vals", {
     }
   }
 })
-
-test_that("Variance algorithm takes longer for smaller tolerance.", {
-  d <- dst_norm(0, 1)
-  t1 <- system.time(algorithm_variance(d, tol = 1e-4))
-  t2 <- system.time(algorithm_variance(d, tol = 1e-8))
-  expect_lt(t1[["elapsed"]], t2[["elapsed"]])
-})
