+++ DESCRIPTION
@@ -26,7 +26,6 @@ Roxygen: list (markdown = TRUE, roclets = c ("namespace", "rd", "srr::srr_stats_
 RoxygenNote: 7.3.2
 Imports: 
     checkmate,
-    cubature,
     ellipsis,
     rlang,
     stats,
