### 2025-04-07 53b92a6
organize the srrstandards; get ready to copy them into the codebase.
