+++ R/eval_cdf.R
@@ -35,10 +35,6 @@
 #' enframe_cdf(
 #'   model1 = d1, model2 = d2, at = 0:4, arg_name = "value"
 #' )
-#' @srrstats {PD3.1} Operations on probability distributions are
-#' contained within separate functions which themselves accept the
-#' names of the distributions as one input parameter. Examples include
-#' the `eval_()` and `enframe_()` families of functions.
 #' @rdname cdf
 #' @export
 eval_cdf <- function(distribution, at) {
