+++ tests/testthat/test-edge_cases-eval_functions.R
@@ -1,18 +1,5 @@
 # Make sure that defined distributions evaluate NA inputs properly.
 # Make sure they evaluate edge cases properly.
-#' @srrstats {G5.8} Edge conditions are tested when evaluating
-#' representations
-#' @srrstats {G5.8a} Zero-length data input outputs 0-length vectors.
-#' An error is thrown if zero-length parameters are input into `dst_*()`
-#' functions.
-#' @srrstats {G5.8b} Data of unsupported types does not pass the checkmate
-#' checks on function inputs.
-#' @srrstats {G5.8c} Data with all-`NA` fields or columns or all identical
-#' fields or columns is no different from having some `NA` fields.
-#' @srrstats {G5.8d} Data outside of the scope of the (quantile) algorithm
-#' is only applicable when the quantile probability is outside of [0, 1],
-#' in which case an error is thrown (due to a check for valid function
-#' inputs)
 test_that("`eval_()` functions handle edge cases properly", {
   for (item in test_distributions) {
     d <- rlang::exec(item$distribution, !!!item$valid[[1]])
