+++ R/eval_quantile.R
@@ -17,19 +17,12 @@
 #' and is a simple idea that has been passed around on the internet
 #' here and there. Tolerance is less than 1e-9, unless the maximum
 #' number of iterations (200) is reached.
-#' @srrstats {G1.1} *Statistical Software should document whether
-#' the algorithm(s) it implements are:* - *The first implementation of
-#' a novel algorithm*; or - *The first implementation within **R** of
-#' an algorithm which has previously been implemented in other languages
-#' or contexts*; or - *An improvement on other implementations of
-#' similar algorithms in **R***.
-#' Done, the algorithm has no specific reference.
-#' @srrstats {PD3.3} *Return objects which include values generated
-#' from optimisation algorithms should include information on
-#' optimisation algorithm and performance, minimally including the name
-#' of the algorithm used, the convergence tolerance, and the number of
-#' iterations.*
-#' - This is specified.
+#'
+#' The algorithm is not new, and is rather simple.
+#' The algorithm works by progressively
+#' cutting an initially wide range in half, moving into the left or right
+#' half depending on where the solution is. I found the idea on Stack
+#' Overflow somewhere, but unfortunately cannot find the location anymore.
 #' @rdname quantile
 #' @export
 eval_quantile <- function(distribution, at) {
