+++ R/srr-stats-standards.R
@@ -1,41 +1,129 @@
 #' @srrstatsVerbose TRUE
 #'
-#' @srrstats {G1.0} Distributions are generic enough to not need a specific reference (e.g., an intro probability book will cover most concepts in distionary), and it appears most other distribution packages on CRAN do not include references.
-#' @srrstats {G1.2} A Life Cycle Statement describing current and anticipated future states of development can be found in the CONTRIBUTING file.
-#' @srrstats {G1.3} Users are referred to general texts in probability to learn more about probability concepts and terminology. Terms specific to the package (like "distributional representation" are defined in vignettes).
-#' @srrstats {G1.4} Software uses [`roxygen2`](https://roxygen2.r-lib.org/) to document all functions.
-#' @srrstats {G1.4a} All internal (non-exported) functions are documented in standard [`roxygen2`](https://roxygen2.r-lib.org/) format, along with a final `@noRd` tag to suppress automatic generation of `.Rd` files.
+#' Here is a list of all complied standards. Where relevant, the standards
+#' are copied over to a relevant location in the code base, either in a
+#' test script or along with the function definitions.
+#'
+#' ## ---- G 1 ----
+#'
+#' @srrstats {G1.0} Distributions are generic enough to not need a specific
+#' reference (e.g., an intro probability book will cover most concepts in
+#' distionary), and it appears most other distribution packages on CRAN do
+#' not include references.
+#'
+#' @srrstats {G1.1} The only relevant algorithm is the quantile algorithm.
+#' It's included in the documentation that this algorithm is not new, and
+#' was found on Stack Overflow somewhere, but unfortunately cannot find
+#' the location anymore.
+#'
+#' @srrstats {G1.2} A Life Cycle Statement describing current and anticipated
+#' future states of development can be found in the CONTRIBUTING file.
+#'
+#' @srrstats {G1.3} Users are referred to general texts in probability to learn
+#' more about probability concepts and terminology. Terms specific to the
+#' package (like "distributional representation" are defined in vignettes).
+#'
+#' @srrstats {G1.4} Software uses [`roxygen2`](https://roxygen2.r-lib.org/) to
+#' document all functions.
+#' @srrstats {G1.4a} All internal (non-exported) functions are documented in
+#' standard [`roxygen2`](https://roxygen2.r-lib.org/) format, along with a
+#' final `@noRd` tag to suppress automatic generation of `.Rd` files.
+#'
+#' ## ---- G 2 ----
+#'
 #' @srrstats {G2.0} Assertions on lengths of inputs (asserting that
 #' inputs expected to be single- or multi-valued) are explicitly
 #' tested for distribution parameters; implicitly through evaluation
 #' functions.
-#' @srrstats {G2.1} Assertions on types of inputs is conducted using the checkmate package for most functions.
-#' @srrstats {G2.2} Prohibiting or restricting submission of multivariate input (i.e., distributions) to parameters expected to be univariate is done using the checkmate package for relevant functions (e.g., `dst_*()` specifications)
-#' @srrstats {G2.3} Univariate character input specifications are asserted using the checkmate package where relevant (e.g., `.vtype` and `.name` in `distribution()`; `arg_name` and `fn_prefix` in `enframe_*()`).
+#' @srrstats {G2.0a} Explicit secondary documentation of expectations on
+#' lengths of inputs have been provided where relevant. See `dst_norm()`
+#' for an example.
+#'
+#' @srrstats {G2.1} Assertions on types of inputs is conducted using the
+#' checkmate package for most functions.
+#' @srrstats {G2.1a} Explicit secondary documentation of expectations on
+#' data types of all vector inputs are provided. See `dst_norm()` for an
+#' example.
+#'
[…]
