### 2025-04-07 0d7a7cf
Complete srrstats standards
