+++ tests/testthat/test-pdq_gev.R
@@ -1,3 +1,9 @@
+#' @srrstats {G3.0} Appropriate tolerances for approximate equality is
+#' adopted in instances of `expect_equal()`. The default is used, except
+#' for instances where comparison can allow a larger tolerance. --> This
+#' srrstats statement is included in all test files that use a different
+#' tolerance in `expect_equal()` than the default.
+
 test_that("GEV negative scale not allowed.", {
   expect_error(dgev(1:10, 3, -4, 1))
   expect_error(pgev(1:10, 3, -4, 1))
