### 2025-04-07 158520a
Add a network diagram connecting distribution properties, and add to vignette.
