+++ doc/specify.Rmd
@@ -131,3 +131,8 @@ eval_property(linear, "mean")
 eval_property(linear, "another_representation", 1:4)
 eval_property(linear, "g")
 ```
+
+
+```{r, echo=FALSE, out.width="100%", fig.align="center"}
+knitr::include_graphics("network_diagram.svg")
+```
