### 2025-04-07 #688 @vincenzocoia
@mpadge I have updated the distionary package to comply with >50% of the SRR standards. The test scripts should also be easier to follow.

> I can, however, reproduce your {autotest} issue - can you please file and issue [over there](https://github.com/ropensci-review-tools/autotest/issues), and we'll investigate further. Thanks

I finally got around to doing this as well, here it is: https://github.com/ropensci-review-tools/autotest/issues/90

> I see a few issues labeled with [bug] that have remained opened 1+ years.

@maurolepore These have been addressed; the remaining Issues are enhancements.

Let me know if anything else needs clarifying at this point.
