### 2025-04-07 #688 @mpadge
That's great @vincenzocoia, thanks for putting in the work there. You can call `@ropensci-review-bot check srr` to confirm compliance here, and then it'll be over to @maurolepore to guide things through to the next stage.
