+++ R/distionary-package.R
@@ -1,3 +1,9 @@
+#' @srrstats {G1.2} A Life Cycle Statement describing current and anticipated
+#' future states of development can be found in the CONTRIBUTING file.
+#' @srrstats {G1.3} Users are referred to general texts in probability to learn
+#' more about probability concepts and terminology (see README). Terms specific
+#' to the package (like "distributional representation" are defined in
+#' vignettes).
 #' @keywords internal
 "_PACKAGE"
 
