+++ R/parameters.R
@@ -28,6 +28,8 @@
 #' parameters(b) <- list(t = 7)
 #' parameters(b)
 #' @rdname parameters
+#' @srrstats {G2.1} Assertions on types of inputs is conducted using the
+#' checkmate package for most functions.
 #' @export
 parameters <- function(distribution) {
   checkmate::assert_class(distribution, "dst")
