+++ tests/testthat/test-edge_cases-eval_functions.R
@@ -1,5 +1,5 @@
 #' @srrstats {G5.8} Edge conditions are tested when evaluating
-#' representations. --> Copied to `test-edge_cases-*.R`.
+#' representations.
 
 # Make sure that defined distributions evaluate NA inputs properly.
 # Make sure they evaluate edge cases properly.
