### 2025-04-09 d7497e4
Fix integerish lower values
