### 2025-04-09 1751cbe
Keep testing srrstats tags in R/ folder bc pkgcheck can't find them in tests/?
