+++ R/srr-stats-standards.R
@@ -1,79 +1,5 @@
 #' @srrstatsVerbose TRUE
 #'
-#' @srrstats {G5.2} Appropriate error behaviour is tested for all
-#' functions explicitly, but warnings are omitted and saved for a future
-#' version. --> Copied to test-edge_cases-builtin_dists.R
-#' @srrstats {G5.2a} While messages produced within R code by `stop()`,
-#' `warning()`, `message()`, or equivalent are not unique (e.g.,
-#' `dst_norm()`, `dst_pois()`, etc. all have the same `length != 1`
-#' error message), they are unique enough to allow the user to debug.
-#' @srrstats {G5.2b} Explicit tests trigger the `stop()` calls in this
-#' version of distionary. --> Copied to test-edge_cases-builtin_dists.R
-#' @srrstats {G5.3} Functions that are expected to return objects containing no
-#' missing (`NA`) or undefined (`NaN`, `Inf`) values are tested either
-#' implicitly (e.g., `is_distribution()` implicitly checks non-NA value) or
-#' explicitly (e.g., `pretty_name()` is never NA).
-#' @srrstats {G5.4} Correctness tests are conducted to test that
-#' statistical algorithms (calculating properties from other distribution
-#' properties) produce expected results to test distributions with set
-#' parameters. See `test-property-*.R` and `test-evaluation-*.R`.
-#' @srrstats {G5.4b} Implementations of existing methods (cdf, density, ...)
-#' are compared against the stats package where possible. Implementations
-#' like the hazard function that are not found in the stats package are
-#' implemented formulaically and verified by comparing to algorithm based
-#' on the representation's definition.
-#' @srrstats {G5.5} Correctness tests are run with four fixed random seeds,
-#' applicable for testing the `realise()` function (same as G5.6b). --> Copied
-#' to `test-realise.R`. See, for example, `test-representation-chf.R` (copied
-#' there).
-#' @srrstats {G5.6} Parameter recovery is relevant when distributional
-#' properties (like quantiles) are computed from other properties (like the
-#' cdf); these are possible to validate by comparing the intrinsic version
-#' of the property against the derived version as if it were absent. See,
-#' for example, `test-representation-density.R` (copied there).
-#' @srrstats {G5.6a} Parameter recovery tests are conducted using the default
-#' tolerance in the `testthat::expect_equal()` function whenever tests of
-#' user-facing outputs are evaluated.
-#' @srrstats {G5.6b} Parameter recovery tests are run with four fixed random
-#' seeds, applicable for testing the `realise()` function. --> Copied to
-#' `test-realise.R`.
-#' @srrstats {G5.7} The only algorithm thus far is the quantile algorithm,
-#' and its performance has been tested to take longer with a smaller
-#' tolerance. --> Copied to `test-representation-quantile.R`.
-#'
-#' @srrstats {G5.8} Edge conditions are tested when evaluating
-#' representations. --> Copied to `test-edge_cases-*.R`.
-#' @srrstats {G5.8a} Zero-length data input outputs 0-length vectors.
-#' An error is thrown if zero-length parameters are input into `dst_*()`
-#' functions.
-#' @srrstats {G5.8b} Data of unsupported types does not pass the checkmate
-#' checks on function inputs.
-#' @srrstats {G5.8c} Data with all-`NA` fields or columns or all identical
-#' fields or columns is no different from having some `NA` fields.
-#' @srrstats {G5.8d} Data outside of the scope of the (quantile) algorithm
-#' is only applicable when the quantile probability is outside of [0, 1],
-#' in which case an error is thrown (due to a check for valid function
-#' inputs).
-#'
-#' @srrstats {PD4.0} The numeric outputs of probability distribution
[…]
