### 2025-04-09 b8e0080
Try fixing broken vignette network diagram
