+++ vignettes/specify.Rmd
@@ -137,5 +137,5 @@ eval_property(linear, "g")
 The current version of `distionary` derives distribution properties from other properties according to the following network. An arrow from one property to another indicates that the first property can be calculated from the second. For example, `quantile` can be calculated from `cdf`, and `mean` can be calculated from `density`.
 
 ```{r, echo=FALSE, out.width="100%", fig.align="center"}
-knitr::include_graphics("network_diagram.svg")
+knitr::include_graphics("./images/network_diagram.svg")
 ```
