### 2025-04-09 #688 @maurolepore
@vincenzocoia and @mpadge please ping me when you think you're ready for me to announce the package to our editors. I'm watching but I'll appreciate a reminder to avoid delays.
