### 2025-04-11 dd101f0
fix links in website, esp. to network diagram
