+++ README.Rmd
@@ -26,14 +26,14 @@ knitr::opts_chunk$set(
 
 With `distionary`, you can:
 
-1. [Specify](https://distionary.probaverse.com/articles/specify) a probability distribution, and
-2. [Evaluate](https://distionary.probaverse.com/articles/evaluate) the probability distribution.
+1. [Specify](./articles/specify.html) a probability distribution, and
+2. [Evaluate](./articles/evaluate.html) the probability distribution.
 
 The main purpose of `distionary` is to implement a distribution object, and to make distribution calculations available even if they are not specified in the distribution. `distionary` powers the wider [`probaverse` ecosystem](https://probaverse.com) for making probability distributions that are representative of your data, not just out-of-the-bag distributions like the Normal, Poisson, etc.
 
 `distionary` makes reference to common terms regarding probability distributions. If you're uneasy with these terms and concepts, most intro books in probability will be a good resource to learn from. As `distionary` develops, more documentation will be made available so that it's more self-contained.
 
-The name "distionary" is a portmanteau of "distribution" and "dictionary". While a dictionary lists and defines words, `distionary` defines distributions and makes a list of common distribution families available. These distributions act as building blocks for the wider probaverse.
+The name "distionary" is a portmanteau of "distribution" and "dictionary". While a dictionary lists and defines words, `distionary` defines distributions and makes a list of common distribution families available. The built-in distributions act as building blocks for the wider probaverse.
 
 ## Installation
 
@@ -142,4 +142,4 @@ To cite package `distionary` in publications use:
 
 ## Code of Conduct
 
-Please note that the distionary project is released with a [Contributor Code of Conduct](https://distionary.probaverse.com/CODE_OF_CONDUCT.html). By contributing to this project, you agree to abide by its terms.
+Please note that the distionary project is released with a [Contributor Code of Conduct](./CODE_OF_CONDUCT.html). By contributing to this project, you agree to abide by its terms.
