+++ doc/specify.Rmd
@@ -79,7 +79,7 @@ You can make your own distribution using the `distribution()` function. Provide 
 - Any distributional representation `*` invoked by `eval_*()` (e.g., `quantile` for `eval_quantile()`).
 - Any property (e.g., `mean` for `mean()`)
 
-See the [Evaluate a Distribution] vignette for more details on these evaluation functions.
+See the [Evaluate a Distribution](./evaluate.html) vignette for more details on these evaluation functions.
 
 For this version of `distionary`, it is mandatory to specify at least `cdf` in order for non-specified properties to be evaluated. For continuous or discrete distributions, `density` or `pmf` may also be needed.
 
@@ -137,5 +137,5 @@ eval_property(linear, "g")
 The current version of `distionary` derives distribution properties from other properties according to the following network. An arrow from one property to another indicates that the first property can be calculated from the second. For example, `quantile` can be calculated from `cdf`, and `mean` can be calculated from `density`.
 
 ```{r, echo=FALSE, out.width="100%", fig.align="center"}
-knitr::include_graphics("./images/network_diagram.svg")
+knitr::include_graphics("network_diagram.png")
 ```
