### 2025-04-21 #688 @vincenzocoia
Hi @maurolepore , wondering if you've been able to find a handling editor for distionary.
