### 2025-04-29 #688 @helske
---

Hi, I'm your handling editor. The package is very interesting and seems to be in code shape. I have just few minor comments at this point:

- Based on the automated checks, seems that you have aimed for 80 character line width limit in the code, but  lintr notes that the are few cases this is exceeded by few characters.
- The articles in the distionary webpage orders the topic as "Evaluate..." and "Specify...", but the Evaluate starts with "This vignette covers the second goal..", so perhaps the order should be switched. I'm not sure how to do this though, as my guess is the order is automatic?
- In the "Specify" vignette, there is one node without any label in the network
- While you do say that the network presents the inner logic of the package, maybe it would be a good to note that these are not equivalent with theoretical connections of these concepts. For example, by first look I started to wonder the following:
  -Why odds -> pmf is separate from all the rest? If you know the pmf you also know cdf? Similarly, if you know density, you should be able to compute cdf. And of course from cdf you should be able to define pmf and pdf.
- Also continuing with the diagram, in my opinion it would be more clear if the arrows were reversed, i.e. pmf "causes" odds etc. But this is more of a matter of taste.
- Why do you use pmf, cdf, and density, and not pmf, cdf, and pdf?

As a more broader comment, if you haven't already though about it, it might be a good idea to check the [`ggdist`](https://mjskay.github.io/ggdist/) package regarding whether its visualization tools could be useful for distionary as well.

None of the things above are in anyway critical, so I will now start looking for reviewers.
