### 2025-05-04 #688 @maurolepore
Dear all my EIC rotation comes to an end so I'm writing a few notes to hand things over to the next EiC.

I'm glad we now have a handling editor and we're seeking reviewers. Thanks for your work!
