### 2025-06-08 d210e43
Add rOpenSci badge
