+++ README.md
@@ -17,6 +17,8 @@ stable](https://img.shields.io/badge/lifecycle-stable-brightgreen.svg)](https://
 [![Project Status: Active – The project has reached a stable, usable
 state and is being actively
 developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
+[![Status at rOpenSci Software Peer
+Review](https://badges.ropensci.org/688_status.svg)](https://github.com/ropensci/software-review/issues/688)
 <!-- badges: end -->
 
 With `distionary`, you can:
