### 2025-06-23 4e0220b
shape=0.5 is closed endpoint for GEV variance
