### 2025-07-15 #688 @dutangc
Please find my review below.

## Topic: Probability distribution

There are numerous packages implementing general or specific probability distributions, see the task view https://cran.r-project.org/view=Distributions .
The distionary package belongs to the general framework and uses S3 object-oriented programming, see `methods(class="dst")`.

## Review of the current literature

Please make a benchmark againts available alternatives, e.g., distribution3, see https://cran.r-project.org/web/views/Distributions.html#miscellaneous.

## Base functionnalities : d, p, q, r

Core d, p, q, r functions are implemented via `eval_pmf`/`eval_density`, `eval_cdf`, `eval_quantile`, `realize`. However, there are two functions for probability density/mass function compared to base R functions.
Unfortunately those functions are not declared as generic functions.

## Very useful non-standard functions

The package allows to compute various useful quantities for a given probability distribution: hazard rate, odds ratio, return level, survival functions.

Moments can be easily computed via numerical integration for continuous distributions.
The first four usual moments are available: `mean`, `variance`, `skewness`, `kurtosis`.
However, there is no function for raw, centered or standardized moments of any order. With too much difficulty, this could be added.

For discrete distributions, there is no function to compute them.
But it is easy. See example below.
```{r}
stdpmf <- function(x, distribution, order, ...)
  (x - mean(distribution))^order/(variance(distribution)^(order/2))*eval_pmf(distribution, at=x)

mydiscrete <- distribution(
  pmf = function(x) {
    d <- rep(0, length(x))
    d[x == 1] <- 1/2
    d[x == 2] <- 1/3
    d[x == 3] <- 1/6
    d
  },
  cdf = function(x) {
    p <- 1/2*(x >= 0) + 1/3*(x >= 1) + 1/6*(x >= 2)
    p
  },
  .vtype = "discrete",
  .name = "My discrete",
  mean=5/3,
  variance=10/3-(5/3)^2
)
x <- 0:100
sum(stdpmf(x, mydiscrete, 1)) #0
sum(stdpmf(x, mydiscrete, 2)) #1
sum(stdpmf(x, mydiscrete, 3))
sum(stdpmf(x, mydiscrete, 4))
```

## Computation and profiling

```{r}
library(microbenchmark)
```

It appears there is a large difference for d, p, q functions between base R functions and the proposed functions. This is strange because base R functions are called in the end. The `do.call` version is still quicker.
However, the implementation of the RNG is correct.

### discrete distribution

```{r}
d1 <- dst_geom(0.6)
x <- 0:100
microbenchmark(eval_pmf(d1, at = x),
               dgeom(x, 0.6),
               do.call(dgeom, c(parameters(d1), list(x)))
               )

microbenchmark(eval_cdf(d1, at = x),
               pgeom(x, 0.6),
               do.call(pgeom, c(parameters(d1), list(x)))
)
microbenchmark(realize(d1, n= 1e3),
               rgeom(n=1e3, 0.6)
)
```

### continuous distribution

```{r}
d2 <- dst_lnorm(1/2, 1/4)
x <- 0:100
microbenchmark(eval_density(d2, at = x),
               dlnorm(x, 1/2, 1/4),
               do.call(dlnorm, c(parameters(d2), list(x)))
)
microbenchmark(eval_cdf(d2, at = x),
               plnorm(x, 1/2, 1/4),
               do.call(plnorm, c(parameters(d2), list(x)))
)
microbenchmark(realize(d2, n= 1e3),
               rlnorm(n=1e3, 1/2, 1/4)
)
```

## Non-standard arguments and defensive programming

An error is raised for non-numeric arguments.
```{r, eval=FALSE}
dnorm("a")
dst_norm("a")
```

Vector arguments are not accepted.
```{r}
dnorm(0, mean=0:10)
try(eval_density(dst_norm(0:10, 1)))
```

## Misc.

The vignettes for beginners are well written. It could be good to have an overall man page for the package available at `?distionary`.

There should be some use case specific to a particular field. Generic functions may suffer a too much flexibility compared to dedicated packages. Typicall in the risk community, risk measures are very useful. It might to incorporate an example of conditional tail expectation, e.g., `CTE` in https://cran.r-project.org/package=actuar.

In other fields, there are probably also very typical examples to add to show the flexibility of the package.
