+++ R/dst_empirical.R
@@ -1,0 +1,103 @@
+#' Empirical Distribution
+#'
+#' An empirical distribution is a non-parametric way to
+#' estimate a distribution using data. By default,
+#' it assigns equal probability to all observations
+#' (this can be overridden with the `weights` argument).
+#' Identical to [dst_finite()] with weights as probabilities,
+#' except weights need not add to 1.
+#'
+#' @param y <`data-masking`> Numeric vector representing the potential
+#' outcomes of the distribution.
+#' @param weights <`data-masking`> Numeric vector of weights corresponding to
+#' to the outcomes `y`. These will be scaled so that they add up to 1.
+#' @param data Optionally, a data frame to compute `y` and `weights` from.
+#' `NULL` if not present (the default).
+#' @param na_action_y,na_action_w What should be done with `NA` entries in
+#' `y` and `w`eights? See details.
+#' @returns A finite distribution. If only one outcome, returns a degenerate
+#' distribution. Returns a Null distribution if `NA` values are present
+#' and `"null"` is specified as an NA action.
+#' @details
+#' `y` and `weights` are recycled to have the same length, but only
+#' if one of them has length 1.
+#'
+#' `na_action_y` and `na_action_w` specify the NA action for `y` and `weights`.
+#' Options are, in order of precedence:
+#'
+#' - `"fail"`: Throw an error in the presence of `NA`.
+#' - `"null"`: Return a Null distribution in the presence of `NA`.
+#' - `"drop"` (the default for `na_action_w`): Remove outcome-weight pairs
+#'   having an `NA` value in the specified vector.
+#'
+#' @seealso [dst_finite()]
+#' @examples
+#' t <- -2:7
+#' dst_empirical(t)
+#'
+#' # Using a data frame
+#' df <- data.frame(time = c(NA, NA, t))
+#' dst_empirical(time * 60, data = df)  # Null, since `NA` in `time`.
+#'
+#' # Drop NA `time` values.
+#' dst_empirical(time * 60, data = df, na_action_y = "drop")
+#'
+#' # Weights explicit. Zero-weight outcomes ("-120") are gone.
+#' df$w <- c(1, 1, 0:9)
+#' dst_empirical(time * 60, w, data = df, na_action_y = "drop")
+#'
+#' # "Null" takes precedence over "drop".
+#' df$w <- c(NA, NA, 0:9)
+#' dst_empirical(time, w, data = df, na_action_w = "null", na_action_y = "drop")
+#' dst_empirical(time, w, data = df, na_action_w = "drop", na_action_y = "null")
+#' @export
+dst_empirical <- function(y,
+                          weights = 1,
+                          data = NULL,
+                          na_action_y = c("null", "drop", "fail"),
+                          na_action_w = c("null", "drop", "fail")) {
+  enquo_y <- rlang::enquo(y)
+  enquo_w <- rlang::enquo(weights)
+  y <- rlang::eval_tidy(enquo_y, data = data)
+  w <- rlang::eval_tidy(enquo_w, data = data)
+  yw <- vctrs::vec_recycle_common(y, w)
+  y <- yw[[1]]
+  w <- yw[[2]]
+  na_action_y <- rlang::arg_match(na_action_y)
+  na_action_w <- rlang::arg_match(na_action_w)
+  na_w <- is.na(w)
+  na_y <- is.na(y)
+  has_na_w <- any(na_w)
+  has_na_y <- any(na_y)
+  # Fail first
+  if (has_na_w && na_action_w == "fail") {
+    stop(
+      "Weights have NA values. You can either fix the problem, or choose ",
[…]
