### 2025-10-09 bcd568f
revert commits so that I can work on a new branch post ROpenSci feedback
