### 2025-10-09 3f6f890
Revert "revert commits so that I can work on a new branch post ROpenSci feedback"
This reverts commit bcd568f276b7ed708748b46e40e59a565d09e1e3.
