+++ NAMESPACE
@@ -15,8 +15,10 @@ export(dst_binom)
 export(dst_cauchy)
 export(dst_chisq)
 export(dst_degenerate)
+export(dst_empirical)
 export(dst_exp)
 export(dst_f)
+export(dst_finite)
 export(dst_gamma)
 export(dst_geom)
 export(dst_gev)
