+++ R/eval_from_network-mean.R
@@ -10,32 +10,10 @@ eval_mean_from_network <- function(distribution, tol = 1e-9, ...) {
   dens <- representation_as_function(distribution, representation = "density")
   integrand <- function(x) x * dens(x)
   r <- range(distribution)
-  # if (r[1] == -Inf && r[2] == Inf) {
-  #   # Cauchy distribution mistakenly comes back with finite mean; break
-  #   # integral into two to solve this issue.
-  #   int1 <- try(
-  #     cubature::hcubature(
-  #       integrand,
-  #       lower = r[1], upper = 0,
-  #       tol = tol,
-  #       ...
-  #     ),
-  #     silent = TRUE
-  #   )
-  #   r[1] <- 0
-  # } else {
-  #   int1 <- 0
-  # }
-  # int <- try(
-  #   int1 + cubature::hcubature(
-  #     integrand,
-  #     lowerLimit = r[1], upperLimit = r[2],
-  #     tol = tol,
-  #     ...
-  #   )$integral,
-  #   silent = TRUE
-  # )
-  int <- try(distionary_integrate(integrand, r[1], r[2], tol = tol, ...), silent = TRUE)
+  int <- try(
+    distionary_integrate(integrand, r[1], r[2], tol = tol, ...),
+    silent = TRUE
+  )
   if (inherits(int, "try-error")) {
     message(
       "Integration routine for numerical computation of mean failed. ",
