+++ R/dst_lp3.R
@@ -68,15 +68,15 @@ dst_lp3 <- function(meanlog, sdlog, skew) {
   cdf_gamma <-
     distribution(
       .parameters = list(meanlog = meanlog, sdlog = sdlog, skew = skew),
-      cdf = \(x) stats::pgamma(
+      cdf = function(x) stats::pgamma(
         log(pmax(0, x)) - shift,
         shape = shape, scale = scale
       ),
-      survival = \(x) stats::pgamma(
+      survival = function(x) stats::pgamma(
         log(pmax(0, x)) - shift,
         shape = shape, scale = scale, lower.tail = FALSE
       ),
-      density = \(x) {
+      density = function(x) {
         res <- stats::dgamma(
           log(pmax(0, x)) - shift,
           shape = shape, scale = scale
@@ -84,10 +84,12 @@ dst_lp3 <- function(meanlog, sdlog, skew) {
         res[x == 0] <- 0
         res
       },
-      quantile = \(p) exp(
+      quantile = function(p) exp(
         stats::qgamma(p, shape = shape, scale = scale) + shift
       ),
-      realise = \(n) exp(stats::rgamma(n, shape = shape, scale = scale) + shift),
+      realise = function(n) exp(
+        stats::rgamma(n, shape = shape, scale = scale) + shift
+      ),
       .name = "Log Pearson Type III",
       .vtype = "continuous"
     )
