+++ R/eval_from_network-skewness.R
@@ -18,32 +18,10 @@ eval_skewness_from_network <- function(distribution, tol = 1e-9, ...) {
   dens <- representation_as_function(distribution, representation = "density")
   integrand <- function(x) ((x - mu) / sigma)^3 * dens(x)
   r <- range(distribution)
-  # if (r[1] == -Inf && r[2] == Inf) {
-  #   # t(3) distribution mistakenly comes back with finite skewness; break
-  #   # integral into two to solve this issue.
-  #   int1 <- try(
-  #     cubature::hcubature(
-  #       integrand,
-  #       lowerLimit = r[1], upperLimit = 0,
-  #       tol = tol,
-  #       ...
-  #     )$integral,
-  #     silent = TRUE
-  #   )
-  #   r[1] <- 0
-  # } else {
-  #   int1 <- 0
-  # }
-  # int <- try(
-  #   int1 + cubature::hcubature(
-  #     integrand,
-  #     lowerLimit = r[1], upperLimit = r[2],
-  #     tol = tol,
-  #     ...
-  #   )$integral,
-  #   silent = TRUE
-  # )
-  int <- try(distionary_integrate(integrand, r[1], r[2], tol = tol, ...), silent = TRUE)
+  int <- try(
+    distionary_integrate(integrand, r[1], r[2], tol = tol, ...),
+    silent = TRUE
+  )
   if (inherits(int, "try-error")) {
     message(
       "Integration routine for numerical computation of skewness failed. ",
