### 2025-10-14 probaverse/distionary#37 @vincenzocoia
Raw, centered, or standardized moments
As per the [ROpenSci review](https://github.com/ropensci/software-review/issues/688):

> The first four usual moments are available: `mean`, `variance`, `skewness`, `kurtosis`.
> However, there is no function for raw, centered or standardized moments of any order. With too much difficulty, this could be added.
