### 2025-10-14 c8c49e0
Move numerical integration failure message to be internal to distionary_integrate().
