+++ R/eval_from_network-mean.R
@@ -10,16 +10,5 @@ eval_mean_from_network <- function(distribution, tol = 1e-9, ...) {
   dens <- representation_as_function(distribution, representation = "density")
   integrand <- function(x) x * dens(x)
   r <- range(distribution)
-  int <- try(
-    distionary_integrate(integrand, r[1], r[2], tol = tol, ...),
-    silent = TRUE
-  )
-  if (inherits(int, "try-error")) {
-    message(
-      "Integration routine for numerical computation of mean failed. ",
-      "This could be because the mean does not exist. Returning NaN."
-    )
-    return(NaN)
-  }
-  int
+  distionary_integrate(integrand, lower = r[1], upper = r[2], tol = tol, ...)
 }
