+++ R/eval_from_network-variance.R
@@ -23,21 +23,5 @@ algorithm_variance <- function(distribution, tol = 1e-9, ...) {
   dens <- representation_as_function(distribution, representation = "density")
   integrand <- function(x) (x - mu)^2 * dens(x)
   r <- range(distribution)
-  int <- try(
-    distionary_integrate(
-      integrand,
-      lower = r[1], upper = r[2],
-      tol = tol,
-      ...
-    ),
-    silent = TRUE
-  )
-  if (inherits(int, "try-error")) {
-    message(
-      "Integration routine for numerical computation of variance failed. ",
-      "This could be because the variance does not exist. Returning NaN."
-    )
-    return(NaN)
-  }
-  int
+  distionary_integrate(integrand, lower = r[1], upper = r[2], tol = tol, ...)
 }
