### 2025-10-14 dca6543
Add more detailed man page.
