### 2025-10-14 fab6a27
Fix warning with non-integer dpois inputs.
