### 2025-10-14 69b78c2
misc ROpenSci improvements
