+++ NAMESPACE
@@ -57,7 +57,6 @@ export(is.distribution)
 export(is_distribution)
 export(kurtosis)
 export(kurtosis_exc)
-export(new_distribution)
 export(parameters)
 export(pgev)
 export(pgpd)
