### 2025-10-14 5894bcc
Oops, CONTRIBUTING was already there in .github/ -- update
