+++ R/mean.R
@@ -19,7 +19,7 @@
 #'
 #' @returns A single numeric.
 #' @examples
-#' a <- dst_gpd(1, 0.5)
+#' a <- dst_gp(1, 0.5)
 #' b <- dst_unif(0, 1)
 #' c <- dst_norm(3, 4)
 #' mean(a)
