+++ man/moments.Rd
@@ -50,7 +50,7 @@ could be highly approximate. An upcoming version of distionary will
 resolve this issue.
 }
 \examples{
-a <- dst_gpd(1, 0.5)
+a <- dst_gp(1, 0.5)
 b <- dst_unif(0, 1)
 c <- dst_norm(3, 4)
 mean(a)
