+++ vignettes/specify.Rmd
@@ -37,7 +37,7 @@ All distribution families found in the `stats` package are made available with `
 | Gamma | `dst_gamma()` | Yes |
 | Geometric | `dst_geom()` | Yes |
 | Generalised Extreme Value (GEV) | `dst_gev()` | No |
-| Generalised Pareto (GPD) | `dst_gpd()` | No |
+| Generalised Pareto (GP) | `dst_gp()` | No |
 | Hypergeometric | `dst_hyper()` | Yes |
 | Log Normal | `dst_lnorm()` | Yes |
 | Log Pearson Type III | `dst_lp3()` | No |
