### 2025-10-15 baff944
split "specify" vignette into two
