+++ vignettes/specify-user-defined.Rmd
@@ -1,0 +1,245 @@
+---
+title: "Specifying Your Own Distribution"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Specifying Your Own Distribution}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+```{r setup}
+library(distionary)
+```
+
+This vignette explains how to create user-defined distributions using `distionary`.
+
+## User-Defined Distributions
+
+You can make your own distribution using the `distribution()` function. Specify distribution properties as name-value pairs, where the name is the distribution property. Distributional _representations_ like the cumulative distribution function (CDF) can be specified as functions.
+
+```{r}
+my_normal <- distribution(
+  density = stats::dnorm,
+  cdf = stats::pnorm,
+  range = c(-Inf, Inf),
+  g = 9.81,
+  another_representation = function(x) x^2,
+  .vtype = "continuous"
+)
+# Inspect
+my_normal
+```
+
+The usual evaluation framework can now be accessed, such as evaluating the CDF.
+
+```{r}
+eval_cdf(my_normal, at = c(0.2, 0.5, 0.7))
+```
+
+Even though the mean has not been specified, it can still be evaluated.
+
+```{r}
+mean(my_normal)
+```
+
+Some of the specified properties are special because they correspond to a property that's known to distionary -- in this example, `density`, `cdf`, and `range`. In general, the special names can be identified as follows:
+
+- The suffix following `eval_` for any distributional representation (e.g., `quantile` for `eval_quantile()`). These should be specified as functions.
+- `realise` or `realize`, which is used to generate random samples from the distribution. This should be a function of the number of observations to draw.
+- The same name as the `distionary` functions used to extract any other property (e.g., `mean` for `mean()`).
+
+See the [Evaluate a Distribution](./evaluate.html) vignette for more details on these evaluation functions and the network of relationships between the understood properties. Note that, in the current version, specifying a `cdf` along with a `density` or `pmf` (probability mass function) is required for evaluating non-explicit properties.
+
+In general, properties can be invoked by the more general function `eval_property()`, which is useful for accessing entries that are not known to `distionary`. For example, the user-defined property `g` can be accessed in this way.
+
+```{r}
+eval_property(my_normal, "g")
+```
+
+To use `eval_property()` to evaluate a function, arguments can be specified in the `...` argument.
+
+```{r}
+eval_property(my_normal, "another_representation", 1:4)
+```
+
+The `eval_property()` function is also useful for programmatically accessing distribution properties.
+
+```{r}
+properties <- c("mean", "variance")
+sapply(properties, \(x) eval_property(my_normal, x))
+```
+
+## Distribution Attributes
+
+Distribution metadata are stored as attributes, and are those arguments prefixed by `.` in the `distribution()` function.
+
+- `.vtype`, as seen in the example starting this vignette, is used to specify the variable type, such as `"discrete"` or `"continuous"`.
+- `.name` allows you to give the distribution a name.
+- `.parameters` allows you to specify values for the distribution's parameters as a list, if applicable. This version of `distionary` does not draw from these values, so it currently serves a similar function as the distribution's name.
+
+```{r}
+my_distribution <- distribution(
+  cdf = pnorm,
+  density = dnorm,
+  .vtype = "continuous",
+  .name = "Special",
+  .parameters = list(theta = 1.7, mat = diag(2), hello = "hi")
+)
+# Inspect
+my_distribution
+```
+
+Retrieve the variable type:
+
+```{r}
+vtype(my_distribution)
+```
+
[…]
