+++ man/range.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/range.r
+% Please edit documentation in R/range.R
 \name{range.dst}
 \alias{range.dst}
 \title{Range of Distribution}
