+++ README.md
@@ -23,15 +23,16 @@ Review](https://badges.ropensci.org/688_status.svg)](https://github.com/ropensci
 
 With `distionary`, you can:
 
-1.  [Specify](./articles/specify.html) a probability distribution, and
+1.  Specify a probability distribution
+    ([built-in](./articles/specify-built-in.html) or [on your
+    own](./articles/specify-user-defined.html)), and
 2.  [Evaluate](./articles/evaluate.html) the probability distribution.
 
 The main purpose of `distionary` is to implement a distribution object,
 and to make distribution calculations available even if they are not
-specified in the distribution. `distionary` powers the wider
-[`probaverse` ecosystem](https://probaverse.com) for making probability
-distributions that are representative of your data, not just
-out-of-the-bag distributions like the Normal, Poisson, etc.
+specified in the distribution. `distionary` provides the building blocks
+of the wider [`probaverse` ecosystem](https://probaverse.com) for making
+representative statistical models.
 
 `distionary` makes reference to common terms regarding probability
 distributions. If you’re uneasy with these terms and concepts, most
@@ -47,19 +48,38 @@ wider probaverse.
 
 ## Statement of Need
 
-In stochastic modeling and risk assessment, distributions should
-accurately reflect your data, rather than relying solely on
-out-of-the-box options like the Normal or Poisson distributions.
-Achieving realistic probability distributions demands a versatile
-workbench where distributions can be manipulated and data can inform
-their features. This is the goal of the broader probaverse ecosystem,
-with `distionary` playing a foundational role.
+When building statistical models, distributions should accurately
+reflect your data, but out-of-the-box options like the Normal or Poisson
+distributions often fall short. Achieving realistic probability
+distributions demands a versatile workbench where distributions can be
+manipulated, and data can inform their features. This is the goal of the
+`probaverse` ecosystem, with `distionary` providing the foundational
+building blocks.
 
-As a core package within `probaverse`, `distionary` defines the
-probability distribution object and enables the evaluation of
-distribution properties, even when not explicitly specified. It is
-useful as a standalone package for users who can specify a distribution
-in some forms but need to evaluate it in others.
+As a core component of the `probaverse`, `distionary` provides the
+fundamental infrastructure for defining probability distribution
+objects. It allows for the evaluation of distribution properties, even
+if they aren’t explicitly specified, offering standalone utility for
+users needing to define a distribution in various forms and evaluate it
+comprehensively.
+
+### Target Audience
+
+Lots of people work with probability distributions. Lots of people don’t
+work with probability distributions but should, because they don’t see
+the value or because distributions are too clumsy to work with under
+existing infrastructure. And, there are lots of people learning about
+probability distributions that would have an easier time if they get to
+“feel” distributions and their multifaceted nature. `distionary` is for
+all of these people.
+
+`distionary` is designed for data scientists, statisticians, and
+researchers who require the flexibility to develop custom statistical
+models. It caters to those in finance, insurance, environmental science,
+and engineering, where nuanced distribution modeling is crucial. Whether
+building complex stochastic models or performing detailed risk
+assessments, `distionary` equips users with the tools needed to explore
+and manipulate probability distributions effectively.
 
 ## Installation
 
@@ -203,21 +223,42 @@ enframe_quantile(linear, at = c(0.2, 0.5, 0.9))
 #> 3   0.9    0.684
 ```
 
-## `distionary` in the context of other packages
+## `distionary` in the Context of Other Packages
 
-Other R packages exist that turn probability distributions into objects
-and allow their evaluation. `distionary` is unique in that provides the
-distribution framework needed to power the wider [`probaverse`
[…]
