### 2025-10-15 2982cc8
Change plot() to default to CDF for Empirical distributions.
