+++ DESCRIPTION
@@ -3,7 +3,7 @@ Title: Create and Evaluate Probability Distributions
 Version: 0.1.0
 Authors@R: c(
     person("Vincenzo", "Coia", , "vincenzo.coia@gmail.com", 
-           role = c("aut", "cre")),
+           role = c("aut", "cre", "cph")),
     person("Amogh", "Joshi", role = "ctb"),
     person("Shuyi", "Tan", role = "ctb"),
     person("Zhipeng", "Zhu", role = "ctb")
