### 2025-11-06 #688 @katrinabrock
## Reviewer Response

I didn't examine the entire diff for possible new issues, but I rechecked the blocking issues from my first review:

- Elements missing from README
- Missing Contrib guide
- Poor arg checking

All these are fixed. Looks like most of the minor issues I raised are fixed as well, but I didn't not go through those item by item.

#### Final approval (post-review)

- [X] **The author has responded to my review and made changes to my satisfaction. I recommend approving this package on the condition that the `development` branch gets merged.**

Estimated hours spent reviewing: 10.5
