### 2025-11-07 #688 @vincenzocoia
Thank you @katrinabrock ! @helske is this package now ready for a NEWS update and submission to CRAN?
