### 2025-11-07 #688 @vincenzocoia
Thanks, everyone!

@helske I wonder if it makes sense to keep the repository under `probaverse/distionary`, because the `probaverse` Organization has other packages that rely on distionary and these make up the entire cohesive probaverse. The same goes for the website: it would be more cohesive as it currently is, under `https://distionary.probaverse.com/`.

If this is OK with you, I have undertaken all final steps and am ready to hit `devtools::submit_cran()`!
